move visualizer files
