+++ src/renderer/features/visualizer/components/audiomotionanalyzer/visualizer-settings-form.module.css
