+++ src/renderer/features/visualizer/components/audiomotionanalyzer/visualizer-settings-modal.tsx
@@ -1,4 +1,4 @@
-import { VisualizerSettingsForm } from '/@/renderer/features/player/components/visualizer-settings-form';
+import { VisualizerSettingsForm } from './visualizer-settings-form';
 import { TextTitle } from '/@/shared/components/text-title/text-title';
 import { Stack } from '/@/shared/components/stack/stack';
