+++ src/renderer/features/visualizer/components/audiomotionanalyzer/visualizer.module.css
