+++ src/renderer/features/visualizer/components/audiomotionanalyzer/visualizer.tsx
