+++ src/renderer/router/app-router.tsx
@@ -2,7 +2,7 @@ import { lazy, Suspense } from 'react';
 import { HashRouter, Route, Routes } from 'react-router';
 
 import { ShuffleAllContextModal } from '/@/renderer/features/player/components/shuffle-all-modal';
-import { VisualizerSettingsContextModal } from '/@/renderer/features/player/components/visualizer-settings-modal';
+import { VisualizerSettingsContextModal } from '../features/visualizer/components/audiomotionanalyzer/visualizer-settings-modal';
 import { AddToPlaylistContextModal } from '/@/renderer/features/playlists/components/add-to-playlist-context-modal';
 import { SaveAndReplaceContextModal } from '/@/renderer/features/playlists/components/save-and-replace-context-modal';
 import { UpdatePlaylistContextModal } from '/@/renderer/features/playlists/components/update-playlist-form';
